refactor(orchestra): relocate findElement-using swipe/scrollUntilVisible + route assertLightMode
Relocates swipeCommand and scrollUntilVisible verbatim from Orchestra into
LegacyExecutionBackend, threading BackendContext through their findElement calls
(reusing the backend's Task 1.2 findElement, not re-copying it) and preserving
loop structure, timeouts, and every maestro.* call byte-for-byte. Routes
AssertLightModeCommand into the backend group, calling the already-copied
assertDarkMode(expected: Boolean) helper from Task 1.5 with expected = false —
AssertDarkModeCommand and AssertLightModeCommand now share the one backend copy.
Extends RouterSeamTest with the three newly-routed types and adds seam tests for
swipe dispatch and both dark/light mode assert outcomes.
